renamed SimHitsConstants.h to HcalSimHitsConstants.h to emphasize that the constants are being defined for Hcal Validation
