+++ Validation/HcalHits/interface/HcalSimHitsConstants.h
